refactor(settings): move plugin settings from main to settings
